raft.h: Hide struct raft_progress implementation
