+++ include/raft.h
@@ -537,19 +537,7 @@ struct raft_fsm
  */
 enum { RAFT_UNAVAILABLE, RAFT_FOLLOWER, RAFT_CANDIDATE, RAFT_LEADER };
 
-/**
- * Used by leaders to keep track of replication progress for each server.
- */
-struct raft_progress
-{
-    unsigned short state;         /* Probe, pipeline or snapshot. */
-    raft_index next_index;        /* Next entry to send. */
-    raft_index match_index;       /* Highest index reported as replicated. */
-    raft_index snapshot_index;    /* Last index of most recent snapshot sent. */
-    raft_time last_send;          /* Timestamp of last AppendEntries RPC. */
-    raft_time snapshot_last_send; /* Timestamp of last InstallSnaphot RPC. */
-    bool recent_recv;             /* A msg was received within election timeout. */
-};
+struct raft_progress;
 
 struct raft; /* Forward declaration. */
