+++ test/integration/test_membership.c
@@ -1,4 +1,5 @@
 #include "../../src/configuration.h"
+#include "../../src/progress.h"
 #include "../lib/cluster.h"
 #include "../lib/runner.h"
