+++ test/integration/test_replication.c
@@ -1,3 +1,4 @@
+#include "../../src/progress.h"
 #include "../lib/cluster.h"
 #include "../lib/runner.h"
